refactor rdb tracing package
